[ignore] Moved get_epg_uuid into utils.
